Donation_tracker/requirements.txt
asgiref==3.8.1
Django==5.0.6
django-phonenumber-field==7.3.0
djangorestframework==3.15.1
phonenumbers==8.13.39
python-dotenv==1.0.1
sqlparse==0.5.0
django-cors-headers==4.6.0
